+++ src/main/java/com/springboot/ecommerce/model/productMeta/ProductMetaService.java
@@ -6,9 +6,9 @@ import java.util.List;
 public interface ProductMetaService {
     void saveProductMeta(ProductMeta productMeta);
 
-    void deleteByProduct(Long productId);
+    void deleteByProduct(Integer productId);
 
-    List<ProductMeta> getAllByProduct(Long productId);
+    List<ProductMeta> getAllByProduct(Integer productId);
 
     void deleteProductMeta(Long productMetaId);
